+++ benches/methods_perf_x1000.rs
@@ -2,7 +2,7 @@ use ttf_parser as ttf;
 
 fn units_per_em(bencher: &mut bencher::Bencher) {
     let font_data = std::fs::read("fonts/SourceSansPro-Regular.ttf").unwrap();
-    let face = ttf::Face::from_slice(&font_data, 0).unwrap();
+    let face = ttf::Face::parse(&font_data, 0).unwrap();
     bencher.iter(|| {
         for _ in 0..1000 {
             bencher::black_box(face.units_per_em());
@@ -12,7 +12,7 @@ fn units_per_em(bencher: &mut bencher::Bencher) {
 
 fn width(bencher: &mut bencher::Bencher) {
     let font_data = std::fs::read("fonts/SourceSansPro-Regular.ttf").unwrap();
-    let face = ttf::Face::from_slice(&font_data, 0).unwrap();
+    let face = ttf::Face::parse(&font_data, 0).unwrap();
     bencher.iter(|| {
         for _ in 0..1000 {
             bencher::black_box(face.width());
@@ -22,7 +22,7 @@ fn width(bencher: &mut bencher::Bencher) {
 
 fn ascender(bencher: &mut bencher::Bencher) {
     let font_data = std::fs::read("fonts/SourceSansPro-Regular.ttf").unwrap();
-    let face = ttf::Face::from_slice(&font_data, 0).unwrap();
+    let face = ttf::Face::parse(&font_data, 0).unwrap();
     bencher.iter(|| {
         for _ in 0..1000 {
             bencher::black_box(face.ascender());
@@ -32,7 +32,7 @@ fn ascender(bencher: &mut bencher::Bencher) {
 
 fn underline_metrics(bencher: &mut bencher::Bencher) {
     let font_data = std::fs::read("fonts/SourceSansPro-Regular.ttf").unwrap();
-    let face = ttf::Face::from_slice(&font_data, 0).unwrap();
+    let face = ttf::Face::parse(&font_data, 0).unwrap();
     bencher.iter(|| {
         for _ in 0..1000 {
             bencher::black_box(face.underline_metrics().unwrap());
@@ -42,7 +42,7 @@ fn underline_metrics(bencher: &mut bencher::Bencher) {
 
 fn strikeout_metrics(bencher: &mut bencher::Bencher) {
     let font_data = std::fs::read("fonts/SourceSansPro-Regular.ttf").unwrap();
-    let face = ttf::Face::from_slice(&font_data, 0).unwrap();
+    let face = ttf::Face::parse(&font_data, 0).unwrap();
     bencher.iter(|| {
         for _ in 0..1000 {
             bencher::black_box(face.strikeout_metrics().unwrap());
@@ -52,7 +52,7 @@ fn strikeout_metrics(bencher: &mut bencher::Bencher) {
 
 fn subscript_metrics(bencher: &mut bencher::Bencher) {
     let font_data = std::fs::read("fonts/SourceSansPro-Regular.ttf").unwrap();
-    let face = ttf::Face::from_slice(&font_data, 0).unwrap();
+    let face = ttf::Face::parse(&font_data, 0).unwrap();
     bencher.iter(|| {
         for _ in 0..1000 {
             bencher::black_box(face.subscript_metrics().unwrap());
@@ -62,7 +62,7 @@ fn subscript_metrics(bencher: &mut bencher::Bencher) {
 
 fn x_height(bencher: &mut bencher::Bencher) {
     let font_data = std::fs::read("fonts/SourceSansPro-Regular.ttf").unwrap();
-    let face = ttf::Face::from_slice(&font_data, 0).unwrap();
+    let face = ttf::Face::parse(&font_data, 0).unwrap();
     bencher.iter(|| {
         for _ in 0..1000 {
             bencher::black_box(face.x_height().unwrap());
@@ -72,7 +72,7 @@ fn x_height(bencher: &mut bencher::Bencher) {
 
 fn glyph_hor_advance(bencher: &mut bencher::Bencher) {
     let font_data = std::fs::read("fonts/SourceSansPro-Regular.ttf").unwrap();
-    let face = ttf::Face::from_slice(&font_data, 0).unwrap();
+    let face = ttf::Face::parse(&font_data, 0).unwrap();
     bencher.iter(|| {
         for _ in 0..1000 {
             bencher::black_box(face.glyph_hor_advance(ttf::GlyphId(2)).unwrap());
@@ -82,7 +82,7 @@ fn glyph_hor_advance(bencher: &mut bencher::Bencher) {
 
 fn glyph_hor_side_bearing(bencher: &mut bencher::Bencher) {
     let font_data = std::fs::read("fonts/SourceSansPro-Regular.ttf").unwrap();
-    let face = ttf::Face::from_slice(&font_data, 0).unwrap();
+    let face = ttf::Face::parse(&font_data, 0).unwrap();
     bencher.iter(|| {
         for _ in 0..1000 {
             bencher::black_box(face.glyph_hor_side_bearing(ttf::GlyphId(2)).unwrap());