+++ testing-tools/ttf-fuzz/src/fuzz-glyph-index.rs
@@ -12,7 +12,7 @@ const CHARS: &[char] = &[
 
 fn main() {
     afl::fuzz!(|data: &[u8]| {
-        if let Some(face) = ttf_parser::Face::from_slice(data, 0) {
+        if let Some(face) = ttf_parser::Face::parse(data, 0) {
             for c in CHARS {
                 let _ = face.glyph_index(*c);
             }