+++ testing-tools/ttf-fuzz/src/fuzz-outline.rs
@@ -3,7 +3,7 @@ extern crate afl;
 
 fn main() {
     afl::fuzz!(|data: &[u8]| {
-        if let Some(face) = ttf_parser::Face::from_slice(data, 0) {
+        if let Some(face) = ttf_parser::Face::parse(data, 0) {
             for id in 0..face.number_of_glyphs() {
                 let _ = face.outline_glyph(ttf_parser::GlyphId(id), &mut Builder(0));
             }