+++ testing-tools/ttf-fuzz/src/fuzz-variable-outline.rs
@@ -3,7 +3,7 @@ extern crate afl;
 
 fn main() {
     afl::fuzz!(|data: &[u8]| {
-        if let Some(mut face) = ttf_parser::Face::from_slice(data, 0) {
+        if let Some(mut face) = ttf_parser::Face::parse(data, 0) {
             if face.set_variation(ttf_parser::Tag::from_bytes(b"wght"), 500.0).is_some() {
                 for id in 0..face.number_of_glyphs() {
                     let _ = face.outline_glyph(ttf_parser::GlyphId(id), &mut Builder(0));